Subnet's have a MapPublicIpOnLaunch attribute which determine whether instances launched in the subnet receive a public ipv4 address on launch.
This PR adds the ability to specify AssociatePublicIpAddress when launching, therefore overriding the default set by the subnet's MapPublicIpOnLaunch attribute.

This is handy if the subnet doesn't by default associate a public ip address and you want one for your instance. Note the subnet also needs an internet gateway for the public ip address to be accessible.