Don't show the ghost positional argument `args`
This is fixed by:
```
wrapper.__signature__ = inspect.signature(func)
```
But in doing so it forces all @arg decorators to reference arguments in
the function  signature. Therefore `--profile` can't exist because it
doesn't match a function argument. It has been renamed to `--config`.
Also, because --config is optional on the CLI, the function signature
needs to provide a default even though it's mandatory for the function.
Because python syntax doesn't allow non-default arguments to follow
default arguments it has been moved to be the last argument.