refactor: use ./venv/ for virtual environment (#199)
use _./venv/_ for the virtual environment so it can be referenced from the pre-commit hooks. This means flake8 and black only need to be installed in one place (the virtualenv) for use by pre-commit and vscode.

Another advantage is dependabot will automatically keep them up to date.

darglint integrates with flake8, so it doesn't need to be run separately from flake8.