feat: [breaking change] change volume size default to AMI size (#292)
Change the volume size default from 100 Gb to the size of the AMI.

Why? This aligns behaviour to that of the AWS console and APIs. Some images, like ubuntu, default to 8Gb which makes sense for small applications or test instances.

To mitigate this change to the defaults, volume_size can now be specified in the config file rather than a hard-coded default.